refactor(validation): extract approval hook runtime
